Adding skeleton #hal.executable.target attribute.
