+++ iree/compiler/Dialect/HAL/IR/HALDialect.cpp
@@ -29,6 +29,28 @@ namespace HAL {
 
 namespace {
 
+// Used for custom printing support.
+struct HALOpAsmInterface : public OpAsmDialectInterface {
+  using OpAsmDialectInterface::OpAsmDialectInterface;
+  /// Hooks for getting an alias identifier alias for a given symbol, that is
+  /// not necessarily a part of this dialect. The identifier is used in place of
+  /// the symbol when printing textual IR. These aliases must not contain `.` or
+  /// end with a numeric digit([0-9]+). Returns success if an alias was
+  /// provided, failure otherwise.
+  LogicalResult getAlias(Attribute attr, raw_ostream &os) const override {
+    if (auto targetAttr = attr.dyn_cast<ExecutableTargetAttr>()) {
+      auto targetName = targetAttr.getTarget().getValue().lower();
+      std::replace(targetName.begin(), targetName.end(), '-', '_');
+      os << "executable_target_" << targetName;
+      return success();
+    } else if (attr.isa<LoweringConfig>()) {
+      os << "config";
+      return success();
+    }
+    return failure();
+  }
+};
+
 // Used to control inlining behavior.
 struct HALInlinerInterface : public DialectInlinerInterface {
   using DialectInlinerInterface::DialectInlinerInterface;
@@ -80,18 +102,6 @@ class HALToVMConversionInterface : public VMConversionDialectInterface {
   }
 };
 
-class LoweringConfigAsmDialectInterface : public OpAsmDialectInterface {
-  using OpAsmDialectInterface::OpAsmDialectInterface;
-
-  LogicalResult getAlias(Attribute attr, raw_ostream &os) const override {
-    if (attr.isa<LoweringConfig>()) {
-      os << "config";
-      return success();
-    }
-    return failure();
-  }
-};
-
 }  // namespace
 
 HALDialect::HALDialect(MLIRContext *context)
@@ -105,125 +115,8 @@ HALDialect::HALDialect(MLIRContext *context)
   addOperations<
 #include "iree/compiler/Dialect/HAL/IR/HALOps.cpp.inc"
       >();
-  addInterfaces<HALInlinerInterface, HALToVMConversionInterface,
-                LoweringConfigAsmDialectInterface>();
-}
-
-//===----------------------------------------------------------------------===//
-// Attribute printing and parsing
-//===----------------------------------------------------------------------===//
-
-Attribute HALDialect::parseAttribute(DialectAsmParser &parser,
-                                     Type type) const {
-  StringRef attrKind;
-  if (failed(parser.parseKeyword(&attrKind))) return {};
-  if (attrKind == BufferConstraintsAttr::getKindName()) {
-    return BufferConstraintsAttr::parse(parser);
-  } else if (attrKind == ByteRangeAttr::getKindName()) {
-    return ByteRangeAttr::parse(parser);
-  } else if (attrKind == DescriptorSetLayoutBindingAttr::getKindName()) {
-    return DescriptorSetLayoutBindingAttr::parse(parser);
-  } else if (attrKind == MatchAlwaysAttr::getKindName()) {
-    return MatchAlwaysAttr::parse(parser);
-  } else if (attrKind == MatchAnyAttr::getKindName()) {
-    return MatchAnyAttr::parse(parser);
-  } else if (attrKind == MatchAllAttr::getKindName()) {
-    return MatchAllAttr::parse(parser);
-  } else if (attrKind == DeviceMatchIDAttr::getKindName()) {
-    return DeviceMatchIDAttr::parse(parser);
-  } else if (attrKind == DeviceMatchMemoryModelAttr::getKindName()) {
-    return DeviceMatchMemoryModelAttr::parse(parser);
-  } else if (attrKind == DeviceMatchFeatureAttr::getKindName()) {
-    return DeviceMatchFeatureAttr::parse(parser);
-  } else if (attrKind == DeviceMatchArchitectureAttr::getKindName()) {
-    return DeviceMatchArchitectureAttr::parse(parser);
-  } else if (attrKind == ExConstantStorageAttr::getKindName()) {
-    return ExConstantStorageAttr::parse(parser);
-  } else if (attrKind == ExPushConstantAttr::getKindName()) {
-    return ExPushConstantAttr::parse(parser);
-  } else if (attrKind == ExOperandBufferAttr::getKindName()) {
-    return ExOperandBufferAttr::parse(parser);
-  } else if (attrKind == ExResultBufferAttr::getKindName()) {
-    return ExResultBufferAttr::parse(parser);
-  }
-  parser.emitError(parser.getNameLoc())
-      << "unknown HAL attribute: " << attrKind;
-  return {};
-}
-
-void HALDialect::printAttribute(Attribute attr, DialectAsmPrinter &p) const {
-  TypeSwitch<Attribute>(attr)
-      .Case<BufferConstraintsAttr, ByteRangeAttr,
-            DescriptorSetLayoutBindingAttr,
-            //
-            ExConstantStorageAttr, ExPushConstantAttr, ExOperandBufferAttr,
-            ExResultBufferAttr,
-            //
-            MatchAlwaysAttr, MatchAnyAttr, MatchAllAttr, DeviceMatchIDAttr,
-            DeviceMatchMemoryModelAttr>(
-          [&](auto typedAttr) { typedAttr.print(p); })
-      .Default(
-          [](Attribute) { llvm_unreachable("unhandled HAL attribute kind"); });
-}
-
-//===----------------------------------------------------------------------===//
-// Type printing and parsing
-//===----------------------------------------------------------------------===//
-
-Type HALDialect::parseType(DialectAsmParser &parser) const {
-  StringRef typeKind;
-  if (parser.parseKeyword(&typeKind)) return {};
-  auto type =
-      llvm::StringSwitch<Type>(typeKind)
-          .Case("allocator", AllocatorType::get(getContext()))
-          .Case("buffer", BufferType::get(getContext()))
-          .Case("buffer_view", BufferViewType::get(getContext()))
-          .Case("command_buffer", CommandBufferType::get(getContext()))
-          .Case("descriptor_set", DescriptorSetType::get(getContext()))
-          .Case("descriptor_set_layout",
-                DescriptorSetLayoutType::get(getContext()))
-          .Case("device", DeviceType::get(getContext()))
-          .Case("event", EventType::get(getContext()))
-          .Case("executable", ExecutableType::get(getContext()))
-          .Case("executable_layout", ExecutableLayoutType::get(getContext()))
-          .Case("ring_buffer", RingBufferType::get(getContext()))
-          .Case("semaphore", SemaphoreType::get(getContext()))
-          .Default(nullptr);
-  if (!type) {
-    parser.emitError(parser.getCurrentLocation())
-        << "unknown HAL type: " << typeKind;
-  }
-  return type;
-}
-
-void HALDialect::printType(Type type, DialectAsmPrinter &p) const {
-  if (type.isa<AllocatorType>()) {
-    p << "allocator";
-  } else if (type.isa<BufferType>()) {
-    p << "buffer";
-  } else if (type.isa<BufferViewType>()) {
-    p << "buffer_view";
-  } else if (type.isa<CommandBufferType>()) {
-    p << "command_buffer";
-  } else if (type.isa<DescriptorSetType>()) {
-    p << "descriptor_set";
-  } else if (type.isa<DescriptorSetLayoutType>()) {
-    p << "descriptor_set_layout";
-  } else if (type.isa<DeviceType>()) {
-    p << "device";
-  } else if (type.isa<EventType>()) {
-    p << "event";
-  } else if (type.isa<ExecutableType>()) {
-    p << "executable";
-  } else if (type.isa<ExecutableLayoutType>()) {
-    p << "executable_layout";
-  } else if (type.isa<RingBufferType>()) {
-    p << "ring_buffer";
-  } else if (type.isa<SemaphoreType>()) {
-    p << "semaphore";
-  } else {
-    llvm_unreachable("unknown HAL type");
-  }
+  addInterfaces<HALInlinerInterface, HALOpAsmInterface,
+                HALToVMConversionInterface>();
 }
 
 //===----------------------------------------------------------------------===//
