+++ iree/compiler/Dialect/HAL/IR/test/BUILD
@@ -31,6 +31,7 @@ iree_lit_test_suite(
             "descriptor_set_ops.mlir",
             "device_ops.mlir",
             "executable_ops.mlir",
+            "executable_targets.mlir",
             "experimental_ops.mlir",
             "interface_ops.mlir",
             "invalid.mlir",
